+++ exporter/elasticexporter/internal/translator/elastic/traces_test.go
@@ -193,66 +193,66 @@ func TestEncodeSpanTruncation(t *testing.T) {
 }
 
 func TestTransactionHTTPRequestURL(t *testing.T) {
-	test := func(t *testing.T, expectedFull string, attrs map[string]pdata.AttributeValue) {
+	test := func(t *testing.T, expectedFull string, attrs map[string]pdata.Value) {
 		transaction := transactionWithAttributes(t, attrs)
 		assert.Equal(t, expectedFull, transaction.Context.Request.URL.Full)
 	}
 	t.Run("scheme_host_target", func(t *testing.T) {
-		test(t, "https://testing.invalid:80/foo?bar", map[string]pdata.AttributeValue{
-			"http.scheme": pdata.NewAttributeValueString("https"),
-			"http.host":   pdata.NewAttributeValueString("testing.invalid:80"),
-			"http.target": pdata.NewAttributeValueString("/foo?bar"),
+		test(t, "https://testing.invalid:80/foo?bar", map[string]pdata.Value{
+			"http.scheme": pdata.NewValueString("https"),
+			"http.host":   pdata.NewValueString("testing.invalid:80"),
+			"http.target": pdata.NewValueString("/foo?bar"),
 		})
 	})
 	t.Run("scheme_servername_nethostport_target", func(t *testing.T) {
-		test(t, "https://testing.invalid:80/foo?bar", map[string]pdata.AttributeValue{
-			"http.scheme":      pdata.NewAttributeValueString("https"),
-			"http.server_name": pdata.NewAttributeValueString("testing.invalid"),
-			"net.host.port":    pdata.NewAttributeValueInt(80),
-			"http.target":      pdata.NewAttributeValueString("/foo?bar"),
+		test(t, "https://testing.invalid:80/foo?bar", map[string]pdata.Value{
+			"http.scheme":      pdata.NewValueString("https"),
+			"http.server_name": pdata.NewValueString("testing.invalid"),
+			"net.host.port":    pdata.NewValueInt(80),
+			"http.target":      pdata.NewValueString("/foo?bar"),
 		})
 	})
 	t.Run("scheme_nethostname_nethostport_target", func(t *testing.T) {
-		test(t, "https://testing.invalid:80/foo?bar", map[string]pdata.AttributeValue{
-			"http.scheme":   pdata.NewAttributeValueString("https"),
-			"net.host.name": pdata.NewAttributeValueString("testing.invalid"),
-			"net.host.port": pdata.NewAttributeValueInt(80),
-			"http.target":   pdata.NewAttributeValueString("/foo?bar"),
+		test(t, "https://testing.invalid:80/foo?bar", map[string]pdata.Value{
+			"http.scheme":   pdata.NewValueString("https"),
+			"net.host.name": pdata.NewValueString("testing.invalid"),
+			"net.host.port": pdata.NewValueInt(80),
+			"http.target":   pdata.NewValueString("/foo?bar"),
 		})
 	})
 	t.Run("http.url", func(t *testing.T) {
 		const httpURL = "https://testing.invalid:80/foo?bar"
-		test(t, httpURL, map[string]pdata.AttributeValue{
-			"http.url": pdata.NewAttributeValueString(httpURL),
+		test(t, httpURL, map[string]pdata.Value{
+			"http.url": pdata.NewValueString(httpURL),
 		})
 	})
 	t.Run("host_no_port", func(t *testing.T) {
-		test(t, "https://testing.invalid/foo?bar", map[string]pdata.AttributeValue{
-			"http.scheme": pdata.NewAttributeValueString("https"),
-			"http.host":   pdata.NewAttributeValueString("testing.invalid"),
-			"http.target": pdata.NewAttributeValueString("/foo?bar"),
+		test(t, "https://testing.invalid/foo?bar", map[string]pdata.Value{
+			"http.scheme": pdata.NewValueString("https"),
+			"http.host":   pdata.NewValueString("testing.invalid"),
+			"http.target": pdata.NewValueString("/foo?bar"),
 		})
 	})
 	t.Run("ipv6_host_no_port", func(t *testing.T) {
-		test(t, "https://[::1]/foo?bar", map[string]pdata.AttributeValue{
-			"http.scheme": pdata.NewAttributeValueString("https"),
-			"http.host":   pdata.NewAttributeValueString("[::1]"),
-			"http.target": pdata.NewAttributeValueString("/foo?bar"),
+		test(t, "https://[::1]/foo?bar", map[string]pdata.Value{
+			"http.scheme": pdata.NewValueString("https"),
+			"http.host":   pdata.NewValueString("[::1]"),
+			"http.target": pdata.NewValueString("/foo?bar"),
 		})
 	})
 
 	// Scheme is set to "http" if it can't be deduced from attributes.
 	t.Run("default_scheme", func(t *testing.T) {
-		test(t, "http://testing.invalid:80/foo?bar", map[string]pdata.AttributeValue{
-			"http.host":   pdata.NewAttributeValueString("testing.invalid:80"),
-			"http.target": pdata.NewAttributeValueString("/foo?bar"),
+		test(t, "http://testing.invalid:80/foo?bar", map[string]pdata.Value{
+			"http.host":   pdata.NewValueString("testing.invalid:80"),
+			"http.target": pdata.NewValueString("/foo?bar"),
 		})
 	})
 }
 
 func TestTransactionHTTPRequestURLInvalid(t *testing.T) {
-	transaction := transactionWithAttributes(t, map[string]pdata.AttributeValue{
-		"http.url": pdata.NewAttributeValueString("0.0.0.0:8081"),
+	transaction := transactionWithAttributes(t, map[string]pdata.Value{
+		"http.url": pdata.NewValueString("0.0.0.0:8081"),
 	})
 	require.NotNil(t, transaction.Context)
 	assert.Nil(t, transaction.Context.Request)
@@ -262,54 +262,54 @@ func TestTransactionHTTPRequestURLInvalid(t *testing.T) {
 }
 
 func TestTransactionHTTPRequestSocketRemoteAddr(t *testing.T) {
-	test := func(t *testing.T, expected string, attrs map[string]pdata.AttributeValue) {
+	test := func(t *testing.T, expected string, attrs map[string]pdata.Value) {
 		transaction := transactionWithAttributes(t, attrs)
 		assert.Equal(t, expected, transaction.Context.Request.Socket.RemoteAddress)
 	}
 	t.Run("net.peer.ip_port", func(t *testing.T) {
-		test(t, "192.168.0.1:1234", map[string]pdata.AttributeValue{
-			"http.url":      pdata.NewAttributeValueString("http://testing.invalid"),
-			"net.peer.ip":   pdata.NewAttributeValueString("192.168.0.1"),
-			"net.peer.port": pdata.NewAttributeValueInt(1234),
+		test(t, "192.168.0.1:1234", map[string]pdata.Value{
+			"http.url":      pdata.NewValueString("http://testing.invalid"),
+			"net.peer.ip":   pdata.NewValueString("192.168.0.1"),
+			"net.peer.port": pdata.NewValueInt(1234),
 		})
 	})
 	t.Run("net.peer.ip", func(t *testing.T) {
-		test(t, "192.168.0.1", map[string]pdata.AttributeValue{
-			"http.url":    pdata.NewAttributeValueString("http://testing.invalid"),
-			"net.peer.ip": pdata.NewAttributeValueString("192.168.0.1"),
+		test(t, "192.168.0.1", map[string]pdata.Value{
+			"http.url":    pdata.NewValueString("http://testing.invalid"),
+			"net.peer.ip": pdata.NewValueString("192.168.0.1"),
 		})
 	})
 	t.Run("http.remote_addr", func(t *testing.T) {
-		test(t, "192.168.0.1:1234", map[string]pdata.AttributeValue{
-			"http.url":         pdata.NewAttributeValueString("http://testing.invalid"),
-			"http.remote_addr": pdata.NewAttributeValueString("192.168.0.1:1234"),
+		test(t, "192.168.0.1:1234", map[string]pdata.Value{
+			"http.url":         pdata.NewValueString("http://testing.invalid"),
+			"http.remote_addr": pdata.NewValueString("192.168.0.1:1234"),
 		})
 	})
 	t.Run("http.remote_addr_no_port", func(t *testing.T) {
-		test(t, "192.168.0.1", map[string]pdata.AttributeValue{
-			"http.url":         pdata.NewAttributeValueString("http://testing.invalid"),
-			"http.remote_addr": pdata.NewAttributeValueString("192.168.0.1"),
+		test(t, "192.168.0.1", map[string]pdata.Value{
+			"http.url":         pdata.NewValueString("http://testing.invalid"),
+			"http.remote_addr": pdata.NewValueString("192.168.0.1"),
 		})
 	})
 }
 
 func TestTransactionHTTPRequestHTTPVersion(t *testing.T) {
-	transaction := transactionWithAttributes(t, map[string]pdata.AttributeValue{
-		"http.flavor": pdata.NewAttributeValueString("1.1"),
+	transaction := transactionWithAttributes(t, map[string]pdata.Value{
+		"http.flavor": pdata.NewValueString("1.1"),
 	})
 	assert.Equal(t, "1.1", transaction.Context.Request.HTTPVersion)
 }
 
 func TestTransactionHTTPRequestHTTPMethod(t *testing.T) {
-	transaction := transactionWithAttributes(t, map[string]pdata.AttributeValue{
-		"http.method": pdata.NewAttributeValueString("PATCH"),
+	transaction := transactionWithAttributes(t, map[string]pdata.Value{
+		"http.method": pdata.NewValueString("PATCH"),
 	})
 	assert.Equal(t, "PATCH", transaction.Context.Request.Method)
 }
 
 func TestTransactionHTTPRequestUserAgent(t *testing.T) {
-	transaction := transactionWithAttributes(t, map[string]pdata.AttributeValue{
-		"http.user_agent": pdata.NewAttributeValueString("Foo/bar (baz)"),
+	transaction := transactionWithAttributes(t, map[string]pdata.Value{
+		"http.user_agent": pdata.NewValueString("Foo/bar (baz)"),
 	})
 	assert.Equal(t, model.Headers{{
 		Key:    "User-Agent",
@@ -318,8 +318,8 @@ func TestTransactionHTTPRequestUserAgent(t *testing.T) {
 }
 
 func TestTransactionHTTPRequestClientIP(t *testing.T) {
-	transaction := transactionWithAttributes(t, map[string]pdata.AttributeValue{
-		"http.client_ip": pdata.NewAttributeValueString("256.257.258.259"),
+	transaction := transactionWithAttributes(t, map[string]pdata.Value{
+		"http.client_ip": pdata.NewValueString("256.257.258.259"),
 	})
 	assert.Equal(t, model.Headers{{
 		Key:    "X-Forwarded-For",
@@ -328,59 +328,59 @@ func TestTransactionHTTPRequestClientIP(t *testing.T) {
 }
 
 func TestTransactionHTTPResponseStatusCode(t *testing.T) {
-	transaction := transactionWithAttributes(t, map[string]pdata.AttributeValue{
-		"http.status_code": pdata.NewAttributeValueInt(200),
+	transaction := transactionWithAttributes(t, map[string]pdata.Value{
+		"http.status_code": pdata.NewValueInt(200),
 	})
 	assert.Equal(t, 200, transaction.Context.Response.StatusCode)
 }
 
 func TestSpanHTTPURL(t *testing.T) {
-	test := func(t *testing.T, expectedURL string, attrs map[string]pdata.AttributeValue) {
+	test := func(t *testing.T, expectedURL string, attrs map[string]pdata.Value) {
 		span := spanWithAttributes(t, attrs)
 		assert.Equal(t, expectedURL, span.Context.HTTP.URL.String())
 	}
 	t.Run("http.url", func(t *testing.T) {
 		const httpURL = "https://testing.invalid:80/foo?bar"
-		test(t, httpURL, map[string]pdata.AttributeValue{
-			"http.url": pdata.NewAttributeValueString(httpURL),
+		test(t, httpURL, map[string]pdata.Value{
+			"http.url": pdata.NewValueString(httpURL),
 		})
 	})
 	t.Run("scheme_host_target", func(t *testing.T) {
-		test(t, "https://testing.invalid:80/foo?bar", map[string]pdata.AttributeValue{
-			"http.scheme": pdata.NewAttributeValueString("https"),
-			"http.host":   pdata.NewAttributeValueString("testing.invalid:80"),
-			"http.target": pdata.NewAttributeValueString("/foo?bar"),
+		test(t, "https://testing.invalid:80/foo?bar", map[string]pdata.Value{
+			"http.scheme": pdata.NewValueString("https"),
+			"http.host":   pdata.NewValueString("testing.invalid:80"),
+			"http.target": pdata.NewValueString("/foo?bar"),
 		})
 	})
 	t.Run("scheme_netpeername_netpeerport_target", func(t *testing.T) {
-		test(t, "https://testing.invalid:80/foo?bar", map[string]pdata.AttributeValue{
-			"http.scheme":   pdata.NewAttributeValueString("https"),
-			"net.peer.name": pdata.NewAttributeValueString("testing.invalid"),
-			"net.peer.ip":   pdata.NewAttributeValueString("::1"), // net.peer.name preferred
-			"net.peer.port": pdata.NewAttributeValueInt(80),
-			"http.target":   pdata.NewAttributeValueString("/foo?bar"),
+		test(t, "https://testing.invalid:80/foo?bar", map[string]pdata.Value{
+			"http.scheme":   pdata.NewValueString("https"),
+			"net.peer.name": pdata.NewValueString("testing.invalid"),
+			"net.peer.ip":   pdata.NewValueString("::1"), // net.peer.name preferred
+			"net.peer.port": pdata.NewValueInt(80),
+			"http.target":   pdata.NewValueString("/foo?bar"),
 		})
 	})
 	t.Run("scheme_netpeerip_netpeerport_target", func(t *testing.T) {
-		test(t, "https://[::1]:80/foo?bar", map[string]pdata.AttributeValue{
-			"http.scheme":   pdata.NewAttributeValueString("https"),
-			"net.peer.ip":   pdata.NewAttributeValueString("::1"),
-			"net.peer.port": pdata.NewAttributeValueInt(80),
-			"http.target":   pdata.NewAttributeValueString("/foo?bar"),
+		test(t, "https://[::1]:80/foo?bar", map[string]pdata.Value{
+			"http.scheme":   pdata.NewValueString("https"),
+			"net.peer.ip":   pdata.NewValueString("::1"),
+			"net.peer.port": pdata.NewValueInt(80),
+			"http.target":   pdata.NewValueString("/foo?bar"),
 		})
 	})
 
 	// Scheme is set to "http" if it can't be deduced from attributes.
 	t.Run("default_scheme", func(t *testing.T) {
-		test(t, "http://testing.invalid:80/foo?bar", map[string]pdata.AttributeValue{
-			"http.host":   pdata.NewAttributeValueString("testing.invalid:80"),
-			"http.target": pdata.NewAttributeValueString("/foo?bar"),
+		test(t, "http://testing.invalid:80/foo?bar", map[string]pdata.Value{
+			"http.host":   pdata.NewValueString("testing.invalid:80"),
+			"http.target": pdata.NewValueString("/foo?bar"),
 		})
 	})
 }
 
 func TestSpanHTTPDestination(t *testing.T) {
-	test := func(t *testing.T, expectedAddr string, expectedPort int, expectedName string, expectedResource string, attrs map[string]pdata.AttributeValue) {
+	test := func(t *testing.T, expectedAddr string, expectedPort int, expectedName string, expectedResource string, attrs map[string]pdata.Value) {
 		span := spanWithAttributes(t, attrs)
 		assert.Equal(t, &model.DestinationSpanContext{
 			Address: expectedAddr,
@@ -393,40 +393,40 @@ func TestSpanHTTPDestination(t *testing.T) {
 		}, span.Context.Destination)
 	}
 	t.Run("url_default_port_specified", func(t *testing.T) {
-		test(t, "testing.invalid", 443, "https://testing.invalid", "testing.invalid:443", map[string]pdata.AttributeValue{
-			"http.url": pdata.NewAttributeValueString("https://testing.invalid:443/foo?bar"),
+		test(t, "testing.invalid", 443, "https://testing.invalid", "testing.invalid:443", map[string]pdata.Value{
+			"http.url": pdata.NewValueString("https://testing.invalid:443/foo?bar"),
 		})
 	})
 	t.Run("url_port_scheme", func(t *testing.T) {
-		test(t, "testing.invalid", 443, "https://testing.invalid", "testing.invalid:443", map[string]pdata.AttributeValue{
-			"http.url": pdata.NewAttributeValueString("https://testing.invalid/foo?bar"),
+		test(t, "testing.invalid", 443, "https://testing.invalid", "testing.invalid:443", map[string]pdata.Value{
+			"http.url": pdata.NewValueString("https://testing.invalid/foo?bar"),
 		})
 	})
 	t.Run("url_non_default_port", func(t *testing.T) {
-		test(t, "testing.invalid", 444, "https://testing.invalid:444", "testing.invalid:444", map[string]pdata.AttributeValue{
-			"http.url": pdata.NewAttributeValueString("https://testing.invalid:444/foo?bar"),
+		test(t, "testing.invalid", 444, "https://testing.invalid:444", "testing.invalid:444", map[string]pdata.Value{
+			"http.url": pdata.NewValueString("https://testing.invalid:444/foo?bar"),
 		})
 	})
 	t.Run("scheme_host_target", func(t *testing.T) {
-		test(t, "testing.invalid", 444, "https://testing.invalid:444", "testing.invalid:444", map[string]pdata.AttributeValue{
-			"http.scheme": pdata.NewAttributeValueString("https"),
-			"http.host":   pdata.NewAttributeValueString("testing.invalid:444"),
-			"http.target": pdata.NewAttributeValueString("/foo?bar"),
+		test(t, "testing.invalid", 444, "https://testing.invalid:444", "testing.invalid:444", map[string]pdata.Value{
+			"http.scheme": pdata.NewValueString("https"),
+			"http.host":   pdata.NewValueString("testing.invalid:444"),
+			"http.target": pdata.NewValueString("/foo?bar"),
 		})
 	})
 	t.Run("scheme_netpeername_nethostport_target", func(t *testing.T) {
-		test(t, "::1", 444, "https://[::1]:444", "[::1]:444", map[string]pdata.AttributeValue{
-			"http.scheme":   pdata.NewAttributeValueString("https"),
-			"net.peer.ip":   pdata.NewAttributeValueString("::1"),
-			"net.peer.port": pdata.NewAttributeValueInt(444),
-			"http.target":   pdata.NewAttributeValueString("/foo?bar"),
+		test(t, "::1", 444, "https://[::1]:444", "[::1]:444", map[string]pdata.Value{
+			"http.scheme":   pdata.NewValueString("https"),
+			"net.peer.ip":   pdata.NewValueString("::1"),
+			"net.peer.port": pdata.NewValueInt(444),
+			"http.target":   pdata.NewValueString("/foo?bar"),
 		})
 	})
 }
 
 func TestSpanHTTPURLInvalid(t *testing.T) {
-	span := spanWithAttributes(t, map[string]pdata.AttributeValue{
-		"http.url": pdata.NewAttributeValueString("0.0.0.0:8081"),
+	span := spanWithAttributes(t, map[string]pdata.Value{
+		"http.url": pdata.NewValueString("0.0.0.0:8081"),
 	})
 	require.NotNil(t, span.Context)
 	assert.Nil(t, span.Context.HTTP)
@@ -436,8 +436,8 @@ func TestSpanHTTPURLInvalid(t *testing.T) {
 }
 
 func TestSpanHTTPStatusCode(t *testing.T) {
-	span := spanWithAttributes(t, map[string]pdata.AttributeValue{
-		"http.status_code": pdata.NewAttributeValueInt(200),
+	span := spanWithAttributes(t, map[string]pdata.Value{
+		"http.status_code": pdata.NewValueInt(200),
 	})
 	assert.Equal(t, 200, span.Context.HTTP.StatusCode)
 }
@@ -445,16 +445,16 @@ func TestSpanHTTPStatusCode(t *testing.T) {
 func TestSpanDatabaseContext(t *testing.T) {
 	// https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/database.md#mysql
 	connectionString := "Server=shopdb.example.com;Database=ShopDb;Uid=billing_user;TableCache=true;UseCompression=True;MinimumPoolSize=10;MaximumPoolSize=50;"
-	span := spanWithAttributes(t, map[string]pdata.AttributeValue{
-		"db.system":            pdata.NewAttributeValueString("mysql"),
-		"db.connection_string": pdata.NewAttributeValueString(connectionString),
-		"db.user":              pdata.NewAttributeValueString("billing_user"),
-		"db.name":              pdata.NewAttributeValueString("ShopDb"),
-		"db.statement":         pdata.NewAttributeValueString("SELECT * FROM orders WHERE order_id = 'o4711'"),
-		"net.peer.name":        pdata.NewAttributeValueString("shopdb.example.com"),
-		"net.peer.ip":          pdata.NewAttributeValueString("192.0.2.12"),
-		"net.peer.port":        pdata.NewAttributeValueInt(3306),
-		"net.transport":        pdata.NewAttributeValueString("IP.TCP"),
+	span := spanWithAttributes(t, map[string]pdata.Value{
+		"db.system":            pdata.NewValueString("mysql"),
+		"db.connection_string": pdata.NewValueString(connectionString),
+		"db.user":              pdata.NewValueString("billing_user"),
+		"db.name":              pdata.NewValueString("ShopDb"),
+		"db.statement":         pdata.NewValueString("SELECT * FROM orders WHERE order_id = 'o4711'"),
+		"net.peer.name":        pdata.NewValueString("shopdb.example.com"),
+		"net.peer.ip":          pdata.NewValueString("192.0.2.12"),
+		"net.peer.port":        pdata.NewValueInt(3306),
+		"net.transport":        pdata.NewValueString("IP.TCP"),
 	})
 
 	assert.Equal(t, "db", span.Type)
@@ -513,7 +513,7 @@ func TestInstrumentationLibrary(t *testing.T) {
 	}, payloads.Transactions[0].Context)
 }
 
-func transactionWithAttributes(t *testing.T, attrs map[string]pdata.AttributeValue) model.Transaction {
+func transactionWithAttributes(t *testing.T, attrs map[string]pdata.Value) model.Transaction {
 	var w fastjson.Writer
 	var recorder transporttest.RecorderTransport
 
@@ -531,7 +531,7 @@ func transactionWithAttributes(t *testing.T, attrs map[string]pdata.AttributeVal
 	return payloads.Transactions[0]
 }
 
-func spanWithAttributes(t *testing.T, attrs map[string]pdata.AttributeValue) model.Span {
+func spanWithAttributes(t *testing.T, attrs map[string]pdata.Value) model.Span {
 	var w fastjson.Writer
 	var recorder transporttest.RecorderTransport