+++ processor/resourcedetectionprocessor/resourcedetection_processor_test.go
@@ -281,7 +281,7 @@ func oCensusResource(res pdata.Resource) *resourcepb.Resource {
 	}
 
 	mp := make(map[string]string, res.Attributes().Len())
-	res.Attributes().Range(func(k string, v pdata.AttributeValue) bool {
+	res.Attributes().Range(func(k string, v pdata.Value) bool {
 		mp[k] = v.StringVal()
 		return true
 	})